refactor: 노선도 패키지 수정, 클래스 이름 수정
