+++ src/main/java/subway/service/SubwayService.java
@@ -7,7 +7,7 @@ import subway.controller.dto.SubwayShortestPathResponse;
 import subway.domain.Path;
 import subway.domain.line.Line;
 import subway.domain.station.Station;
-import subway.domain.subway.DijkstraSubwayRouteMap;
+import subway.domain.subway.routeMap.DijkstraRouteMap;
 import subway.domain.subway.Subway;
 import subway.domain.subway.billing_policy.BillingPolicyByDistance;
 import subway.domain.subway.billing_policy.Fare;
@@ -31,7 +31,7 @@ public class SubwayService {
         final Station sourceStation = stationRepository.findById(sourceStationId);
         final Station destinationStation = stationRepository.findById(destinationStationId);
 
-        final Subway subway = new Subway(new DijkstraSubwayRouteMap(lines), new BillingPolicyByDistance());
+        final Subway subway = new Subway(new DijkstraRouteMap(lines), new BillingPolicyByDistance());
 
         final Path shortestPath = subway.findShortestPath(sourceStation, destinationStation);
         final Fare fare = subway.calculateFare(shortestPath);
